harness: round-7 — merge state, fold claim_guard into providers, extract evolve
Seven files, one object. Round 6 cut tools.py (ghost module); round 7
continues the work by asking which projections were actually the same
primitive at different horizons.

  session_store.py + live_snapshot.py  -> state.py
    Both answered 'where are we right now' — one at conversation
    granularity, one at repo granularity. Same primitive, two horizons.

  claim_guard.py  -> providers.check_claim
    Friction belongs with the boundary it guards. Numeric-claim
    evidence check is a property of the model-output layer, not
    its own module.

  mcp.py (evolve cycle)  -> evolve.py
    MCP serve loop and the nightly self-reflection cycle have
    different lifecycles (long-running server vs one-shot cron).
    Separating them makes each clearer.

Functional changes: none. Every import site rewired. Doctrine block in
__init__.py (_HARNESS_STRATEGY) updated to name the actual seven
modules — same round-6 principle of keeping the description aligned
with what is on disk, since Nemotron reads this file during the
nightly evolve cycle.

Tests: 139 passed, 0 failed, 0 errors. Round-6 baseline preserved.
