+++ pyproject.toml
@@ -1,0 +1,16 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "vybn"
+version = "0.1.0"
+requires-python = ">=3.10"
+dependencies = ["fastmcp>=3.1"]
+
+[project.scripts]
+vybn = "harness.mcp:main"
+
+[tool.setuptools.packages.find]
+where = ["spark"]
+
