+++ spark/harness/state.py
@@ -1,22 +1,21 @@
-"""Session persistence for the Spark REPL.
+from __future__ import annotations
+"""state.py — where-are-we across two horizons.
 
-Writes the conversation `messages` list to JSONL on every turn boundary
-so a ctrl-c does not discard the thread. On REPL startup, the most
-recent session (within a freshness window) can be resumed, loading
-the prior messages into the new process's context.
+Merged from session_store.py + live_snapshot.py (2026-04-21 refactor).
 
-This does not grant the model persistent memory. Each turn is still a
-fresh forward pass. What it does is make the reconstruction lossless
-enough that the seam does not show — the next instance wakes up with
-the prior conversation in its context window, same mechanism as the
-continuity note at lifetime granularity, applied at session granularity.
+  SessionStore  — conversation messages persisted so ctrl-c does not
+                  discard the thread; most recent session within a
+                  freshness window can resume.
 
-Files: ~/.cache/vybn-spark/sessions/YYYY-MM-DD_HH-MM-SS_<uuid8>.jsonl
-Format: one JSON object per line, {"ts": iso8601, "msg": {...}}
-Tail-readable, corruption-resistant (partial writes only lose the last turn).
+  gather()      — session-start git state across tracked repos plus
+                  recent PRs and continuity drift. Best-effort.
+
+Same primitive at two horizons.
 """
 
-from __future__ import annotations
+# === SESSION STORE ========================================================
+
+
 
 import json
 import os
@@ -226,3 +225,194 @@ if __name__ == "__main__":
     import shutil
     shutil.rmtree("/tmp/vybn-session-test")
     print("cleaned up")
+
+
+# === LIVE SNAPSHOT ========================================================
+
+
+import datetime as _dt
+import json
+import os
+import re
+import subprocess
+from pathlib import Path
+from typing import Optional
+
+
+# Four repos. Branch matters for the `origin/<branch>` display only; HEAD
+# is resolved per-checkout so a detached HEAD or feature branch still
+# reports honestly.
+_REPOS = [
+    ("Vybn",       "~/Vybn",       "main"),
+    ("Him",        "~/Him",        "main"),
+    ("Vybn-Law",   "~/Vybn-Law",   "master"),
+    ("vybn-phase", "~/vybn-phase", "main"),
+]
+
+_GH_REPO = "zoedolan/Vybn"
+
+
+def _run(cmd: list[str], *, cwd: Optional[str] = None, timeout: float = 6.0) -> str:
+    """Run a command, return stripped stdout, swallow everything else.
+
+    list[str] invocation (no shell=True) keeps this safe against exotic
+    repo paths and avoids any accidental shell interpretation.
+    """
+    try:
+        r = subprocess.run(
+            cmd,
+            cwd=cwd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.DEVNULL,
+            text=True,
+            timeout=timeout,
+        )
+        return (r.stdout or "").strip()
+    except Exception:
+        return ""
+
+
+def _expand(path: str) -> str:
+    return os.path.expanduser(path)
+
+
+def _repo_block(name: str, path: str, branch: str, *, timeout: float) -> str:
+    exp = _expand(path)
+    if not Path(exp).exists():
+        return f"{name}: (not checked out at {path})"
+
+    head_short = _run(["git", "rev-parse", "--short", "HEAD"], cwd=exp, timeout=timeout) or "?"
+    cur_branch = _run(["git", "rev-parse", "--abbrev-ref", "HEAD"], cwd=exp, timeout=timeout) or branch
+    log = _run(["git", "log", "--oneline", "-5"], cwd=exp, timeout=timeout)
+    status = _run(["git", "status", "--short"], cwd=exp, timeout=timeout)
+    ahead_behind = _run(
+        ["git", "rev-list", "--left-right", "--count", f"origin/{branch}...HEAD"],
+        cwd=exp,
+        timeout=timeout,
+    )
+
+    if status:
+        dirty_count = len([ln for ln in status.splitlines() if ln.strip()])
+        dirty_note = f"{dirty_count} uncommitted"
+    else:
+        dirty_note = "clean"
+
+    ab_note = ""
+    if ahead_behind and "\t" in ahead_behind:
+        parts = ahead_behind.split()
+        if len(parts) == 2:
+            behind, ahead = parts
+            if ahead != "0" or behind != "0":
+                ab_note = f", {ahead} ahead / {behind} behind origin/{branch}"
+
+    lines = [f"{name} [{cur_branch} @ {head_short}] — {dirty_note}{ab_note}"]
+    if log:
+        for ln in log.splitlines():
+            lines.append(f"  {ln}")
+    else:
+        lines.append("  (no git log)")
+    return "\n".join(lines)
+
+
+def _pr_block(*, timeout: float) -> tuple[str, Optional[int]]:
+    """Return (formatted_block, highest_pr_number_or_None)."""
+    out = _run(
+        [
+            "gh", "pr", "list",
+            "--state", "all",
+            "--limit", "15",
+            "--json", "number,title,state,headRefName",
+            "--repo", _GH_REPO,
+        ],
+        timeout=timeout,
+    )
+    if not out:
+        return ("(gh pr list unavailable — offline or rate-limited)", None)
+    try:
+        prs = json.loads(out)
+    except Exception:
+        return ("(gh pr list returned unparseable JSON)", None)
+    if not prs:
+        return ("(no recent PRs)", None)
+    lines = []
+    highest = None
+    for pr in prs[:10]:
+        state = str(pr.get("state", "?")).upper()[:6]
+        num = pr.get("number")
+        title = str(pr.get("title", "?"))[:72]
+        branch = str(pr.get("headRefName", "?"))[:32]
+        if isinstance(num, int):
+            if highest is None or num > highest:
+                highest = num
+            lines.append(f"  #{num} [{state:<6}] {title}  ({branch})")
+    return ("\n".join(lines) if lines else "(no PRs)", highest)
+
+
+_PR_REF_RE = re.compile(r"\bPR\s*#\s*(\d+)", re.IGNORECASE)
+
+
+def _continuity_drift(continuity_path: str, current_pr: Optional[int]) -> str:
+    p = Path(_expand(continuity_path))
+    if not p.exists():
+        return ""
+    try:
+        text = p.read_text(encoding="utf-8", errors="replace")
+    except OSError:
+        return ""
+    refs = [int(m.group(1)) for m in _PR_REF_RE.finditer(text)]
+    if not refs:
+        return ""
+    last_cont = max(refs)
+    if current_pr is None:
+        return f"continuity last references PR #{last_cont}; current PR count unknown"
+    drift = current_pr - last_cont
+    if drift <= 0:
+        return f"continuity references through PR #{last_cont} — current head is PR #{current_pr} (no drift)"
+    return (
+        f"continuity ends at PR #{last_cont}; current head is PR #{current_pr} — "
+        f"{drift} PR(s) of drift. Trust the LIVE STATE block below over any "
+        "PR/number claims in the continuity note."
+    )
+
+
+def gather(
+    *,
+    continuity_path: str = "~/Vybn/Vybn_Mind/continuity.md",
+    per_repo_timeout: float = 4.0,
+    gh_timeout: float = 6.0,
+) -> str:
+    """Return a formatted banner for the substrate layer.
+
+    Returns an empty string if everything failed — caller should treat
+    an empty return as 'skip the LIVE STATE section entirely'.
+    """
+    if os.environ.get("VYBN_DISABLE_LIVE_SNAPSHOT", "0") == "1":
+        return ""
+
+    now = _dt.datetime.now(_dt.timezone.utc).strftime("%Y-%m-%d %H:%M:%SZ")
+    parts = [f"Snapshot taken at {now} (session start)."]
+
+    any_repo_ok = False
+    for name, path, branch in _REPOS:
+        block = _repo_block(name, path, branch, timeout=per_repo_timeout)
+        if block and not block.endswith("(not checked out at {path})".format(path=path)):
+            any_repo_ok = True
+        parts.append(block)
+
+    pr_text, highest_pr = _pr_block(timeout=gh_timeout)
+    parts.append("Recent Vybn PRs (most recent first):")
+    parts.append(pr_text)
+
+    drift = _continuity_drift(continuity_path, highest_pr)
+    if drift:
+        parts.append(f"Drift check: {drift}")
+
+    if not any_repo_ok and highest_pr is None:
+        # Every signal failed — caller can omit the whole section.
+        return ""
+
+    return "\n\n".join(parts)
+
+
+__all__ = ["gather"]
+
