+++ spark/tests/test_claim_guard.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 
-from harness.claim_guard import check  # noqa: E402
+from harness.providers import check_claim as check  # noqa: E402
 
 
 def test_fabricated_numbers_flagged():
